src/main/java/br/com/luiz/pontointeligente/utils/PasswordUtil.java
package br.com.luiz.pontointeligente.utils;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

public class PasswordUtil {

    private static final Logger log = LoggerFactory.getLogger(PasswordUtil.class);

    public static String gerarBCrypt(String senha){
        if(senha == null){
            return senha;
        }

        log.info("Gerando senha com o BCrypt");
        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
        return bCryptPasswordEncoder.encode(senha);
    }
}
